Testing float typecasting.

+++ code/Model/Session.php
@@ -78,7 +78,7 @@ class Cm_RedisSession_Model_Session extends Mage_Core_Model_Mysql4_Session
     public function __construct()
     {
         if ($this->_logLevel >= 7) {
-            $this->_timeStart = microtime(true);
+            $this->_timeStart = (float) microtime(true);
         }
         $host = (string)   (Mage::getConfig()->getNode(self::XML_PATH_HOST) ?: '127.0.0.1');
         $port = (int)      (Mage::getConfig()->getNode(self::XML_PATH_PORT) ?: '6379');
@@ -109,7 +109,7 @@ class Cm_RedisSession_Model_Session extends Mage_Core_Model_Mysql4_Session
                     get_class($this),
                     $host,
                     $port,
-                    (microtime(true) - $this->_timeStart)
+                    ((float) microtime(true) - $this->_timeStart)
                 ),
                 Zend_Log::DEBUG, self::LOG_FILE
             );
